+++ test/Microsoft.AspNetCore.JsonPatch.Test/Internal/ListAdapterTest.cs
