+++ test/Microsoft.AspNetCore.JsonPatch.Test/Internal/ObjectVisitorTest.cs
